Update wording for the guide to creating a pull request.
The button that starts the pull request process is displayed on both the
original and forked repositories, so this updates that paragraph.
Also, if the link is not displayed, you have to go to the forked repository
to initiate a one by clicking 'pull requests' in the right-side menu
instead of doing it from the original.  This commit updates the list items
with those steps too.

And lastly, I only knew how to try and contribute these changes because I
completed the challenges earlier today.  THANK YOU for this guide!